pyproject.toml
[project]
name = "this-studio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=0.27.0",
    "einops>=0.8.2",
    "huggingface-hub>=0.21.0",
    "networkx>=3.2",
    "numpy>=1.24.0",
    "openai>=1.12.0",
    "opencv-python-headless>=4.9.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0",
    "supervision>=0.21.0",
    "tokenizers>=0.15.0",
    "torchvision>=0.24.1",
    "tqdm>=4.66.0",
    "transformers>=4.48.0",
    "vllm>=0.8.0",
]
